[Document] Editable Dialog Box - Add Area Support - #8240
